Criação de alerts e jumbotron, criação de abas na view de cada area para um segunda visualização

+++ platforms/android/assets/www/js/services/lectureseventsAPIservices.js
@@ -1,6 +1,6 @@
 angular.module('isgh.lectureseventsAPIservices', ['isgh.dbAPIservices'])
 
-  .factory('LecturesEvents', function ($q, $http, Constant, DB, $cordovaSQLite, $ionicPopup) {
+  .factory('LecturesEvents', function ($q, $state, $http, Constant, DB, $cordovaSQLite) {
     
     var db = DB;
     var table = Constant.database.tables.lecturesevents;
@@ -14,7 +14,7 @@ angular.module('isgh.lectureseventsAPIservices', ['isgh.dbAPIservices'])
       $http.get(Constant.url_wsapp + 'site/?func=letsAll').then(function (response) {
         deferred.resolve(response);
       }, function (erro) {
-        deferred.reject("Falha na conexão");
+        deferred.reject("Sem conexão com a Internet");
       });
 
       return deferred.promise;
@@ -22,26 +22,34 @@ angular.module('isgh.lectureseventsAPIservices', ['isgh.dbAPIservices'])
     
     // INSERT ROWS IN TABLE
     var _populate = function (refresh) {
-
+      var deferred = $q.defer();
       var refresh = typeof refresh !== 'undefined' ? true : false;
 
-      if (refresh) {
-        db.dropTable(table);
-        db.createTable(table);
-      }
 
-      return _lectureseventsWSget().then(function (response) {
-        angular.forEach(response.data, function (obj) {
-          var query = "INSERT INTO " + table.name + " (" + columns.join(",") + ") values (" + fields.join(",") + ")";
-          db.query(query, [obj.id, obj.title, obj.image, obj.thumbnail, obj.location, obj.location_alias, obj.register_opened, obj.register_closed, obj.event_closed, obj.filename, obj.form_date_up, obj.form_date_down, obj.form_workload, obj.form_location, obj.form_speaker, obj.form_audience, obj.form_investment, obj.form_content_1, obj.form_content_2, obj.form_content_3, obj.form_content_4, obj.form_link, obj.register_link, obj.status]);
-        });
+      _lectureseventsWSget().then(function (response) {
+
+        if (response.data.length > 0) { 
+          if (refresh) {
+            db.dropTable(table);
+            db.createTable(table);
+          }
+          
+          angular.forEach(response.data, function (obj) {
+            var query = "INSERT INTO " + table.name + " (" + columns.join(",") + ") values (" + fields.join(",") + ")";
+            db.query(query, [obj.id, obj.title, obj.image, obj.thumbnail, obj.location, obj.location_alias, obj.date, obj.filename, obj.form_date_up, obj.form_date_down, obj.form_workload, obj.form_location, obj.form_speaker, obj.form_audience, obj.form_investment, obj.form_content_1, obj.form_content_2, obj.form_content_3, obj.form_content_4, obj.form_link, obj.register_link, obj.register_planning, obj.status]);
+          });
+          
+          deferred.resolve(_all());
+          
+        } else {
+          deferred.reject("Restabelecendo conexão perdida com ISGH");
+        }
+
       }, function (erro) {
-        $ionicPopup.alert({
-          title: erro,
-          content: "Verifique se você está conectado à internet e tente novamente."
-        });
+        deferred.reject(erro);
       });
-
+      
+      return deferred.promise;
     }
     
     //SELECT ALL
@@ -49,6 +57,8 @@ angular.module('isgh.lectureseventsAPIservices', ['isgh.dbAPIservices'])
       var query = "SELECT * FROM "+table.name+" ORDER BY id DESC";
       return db.query(query).then(function (result) {
         return db.fetchAll(result);
+      }, function (erro) {
+        console.log(erro);
       });
     };
     
@@ -56,12 +66,11 @@ angular.module('isgh.lectureseventsAPIservices', ['isgh.dbAPIservices'])
       var query = "SELECT * FROM " + table.name + " WHERE id = " + id;
       return db.query(query).then(function (result) {
         return db.fetch(result);
+      }, function (erro) {
+        console.log(erro);
       });
     }
-    // var _get = function (id) {
-    //   return $http.get(Constant.url_appserver + 'site/?func=letsById&id='+id);
-    // }
-    
+        
     return {
       lectureseventsWSget: _lectureseventsWSget,
       populate: _populate,
